Now checks for short buffers (less then CITP header)
This is because the CITP stream is not always aligned with the start of
the packet and unless the Message Length field is in the buffer CITP is
unable to figure out how much data is requires.

This test should now always get enough data to get the full CITP header.